requirements.txt
Django==4.2.7
asgiref==3.7.2
sqlparse==0.4.4
tzdata==2023.3
pycryptodome==3.19.0